+++ src/routes/(app)/eventos/+page.server.ts
@@ -2,21 +2,19 @@ import {
   getSanityServerClient,
   overlayDrafts,
 } from "$lib/config/sanity/client";
-import { allEventsQuery, resultsQuery } from "$lib/config/sanity/queries";
+import { resultsQuery } from "$lib/config/sanity/queries";
 import { error } from "@sveltejs/kit";
 import type { PageServerLoad } from "./$types";
 
 // export const prerender = 'auto';
 export const load: PageServerLoad = async ({ parent, params }) => {
-  const events = await getSanityServerClient(false).fetch(allEventsQuery);
   const results = await getSanityServerClient(false).fetch(resultsQuery);
 
-  if (!events) {
+  if (!results) {
     throw error(500, "Event not found");
   }
 
   return {
-    events,
     results,
   };
 };