+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Add multiples events payment process
+
 ## [0.0.8] - 2023-12-29
 
 - Added disclaimer